fix: compile and clippy fixes for rustc 1.98
rustc 1.98 tightened const-generic checking on type aliases and clippy
grew the manual_div_ceil / unnecessary_cast lints, breaking the stable
CI leg:

- PolyRingVector/PolyRingMatrix aliases declared DEGREE_BOUND as u64
  while PolyRing requires usize; align them to usize
- replace the four hand-rolled (x+7)/8 div_ceil8 bodies with
  usize::div_ceil (stable since 1.73, fine under the 1.85 MSRV)
- drop redundant `bits as usize` casts in mldsa encoding

Verified locally on both CI toolchains (1.98.1 and 1.85.0):
fmt/clippy/build/test/bench all green, 479 tests passing.

+++ crates/algebra/src/matrix/vector.rs
@@ -18,7 +18,7 @@ use rayon::prelude::*;
 pub type RingVector<R> = GenericVector<R>;
 
 /// A vector over a polynomial ring R[x]/(x^d+1)
-pub type PolyRingVector<R, const DEGREE_BOUND: u64> = GenericVector<PolyRing<R, DEGREE_BOUND>>;
+pub type PolyRingVector<R, const DEGREE_BOUND: usize> = GenericVector<PolyRing<R, DEGREE_BOUND>>;
 
 /// Threshold for parallel vector operations
 #[cfg(feature = "parallel")]
